Refactored and fixed minor issues
